Address in gist comments

+++ Octokit/Helpers/ApiUrls.cs
@@ -1313,7 +1313,7 @@ namespace Octokit
         /// </summary>
         /// <param name="gistId">The id of the gist</param>
         /// <param name="commentId">The id of the comment</param>
-        public static Uri GistComment(string gistId, int commentId)
+        public static Uri GistComment(string gistId, long commentId)
         {
             return "gists/{0}/comments/{1}".FormatUri(gistId, commentId);
         }
